Added sequential command group for coral intake
Sequential command group for coral intake was added. Methods used for PID motor movement was changed slightly to be more compatible with the sequential commands. With the addition of more commands, as needed for the sequential command, folders were made for different groups of commands, which github sees as adding and deleting the same thing from different locations.